Add ImageAndTextSubject and shared TextSubject models, refactor SingleTextSubject
