+++ packages/lib-classifier/src/store/SubjectStore/ImageAndTextSubject/index.js
@@ -1,0 +1,1 @@
+export { default } from './ImageAndTextSubject'
