+++ packages/lib-classifier/src/store/SubjectStore/shared/index.js
@@ -1,0 +1,1 @@
+export { default } from './TextSubject.js'
